backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-allauth==0.61.1
dj-rest-auth[with_social]==5.0.2
djangorestframework-simplejwt==5.3.1
google-auth==2.27.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
